rename scripts to reflect move to qcli
the rational for naming of these scripts is that we'll eventually move qcli to a sub-command framework (like git) where there is one main command (qcli) and you pass arguments to it to execute different sub-commands. so eventually i see creating a new qcli script with something like the following:

qcli make_script <some arguments>

so, naming the scripts this way will allow the description of the interface change to essentially be "drop the first underscore from the command".