some edits to clean this code up
it is not currently working to generate rst from qcli/scripts, but it wasn't before these edits either. so, trying to figure out how to proceed here - this script needs to be completely re-written anyway, so i'm considering just leaving it broken for the short-term (at least for this commit).